% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_metrics.R
\name{calc_fcr}
\alias{calc_fcr}
\title{Calculate Feed Conversion Ratio (FCR)}
\usage{
calc_fcr(feed_input, biomass_gain)
}
\arguments{
\item{feed_input}{Numeric. Total dry weight of feed.}

\item{biomass_gain}{Numeric. Total biomass gain.}
}
\value{
Numeric FCR value.
}
\description{
Calculate Feed Conversion Ratio (FCR)
}
