% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_metrics.R
\name{calc_sgr}
\alias{calc_sgr}
\title{Calculate Specific Growth Rate (SGR)}
\usage{
calc_sgr(final_size, initial_size, days)
}
\arguments{
\item{final_size}{Numeric. Final weight or length.}

\item{initial_size}{Numeric. Initial weight or length.}

\item{days}{Numeric. Duration of growth period.}
}
\value{
Numeric SGR value.
}
\description{
Calculate Specific Growth Rate (SGR)
}
