% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_metrics.R
\name{calc_tgc}
\alias{calc_tgc}
\title{Calculate Thermal-Unit Growth Coefficient (TGC)}
\usage{
calc_tgc(final_size, initial_size, temp, days)
}
\arguments{
\item{final_size}{Numeric. Final size.}

\item{initial_size}{Numeric. Initial size.}

\item{temp}{Numeric vector of daily temperatures.}

\item{days}{Numeric. Number of days.}
}
\value{
Numeric TGC value.
}
\description{
Calculate Thermal-Unit Growth Coefficient (TGC)
}
