% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feeding.R
\name{plot_feeding_schedule}
\alias{plot_feeding_schedule}
\title{Visualize Feeding Schedule}
\usage{
plot_feeding_schedule(tidy_data)
}
\arguments{
\item{tidy_data}{Output from read_hatchery_data()}
}
\value{
A ggplot object
}
\description{
Visualize Feeding Schedule
}
