% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_survival_km}
\alias{plot_survival_km}
\title{Plot Kaplan-Meier Survival Curve}
\usage{
plot_survival_km(data, time_col, status_col, group_col)
}
\arguments{
\item{data}{Data frame with time and status columns.}

\item{time_col}{String name of time column (e.g., "dph").}

\item{status_col}{String name of status column (1=dead, 0=alive).}

\item{group_col}{String name of grouping column.}
}
\description{
Plot Kaplan-Meier Survival Curve
}
