% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{potteri_larvae}
\alias{potteri_larvae}
\title{Larval Rearing Data for Centropyge potteri}
\format{
A data frame with rows and variables:
\describe{
  \item{tank_id}{Unique identifier for the rearing vessel}
  \item{dph}{Days Post Hatch}
  \item{protocol}{Experimental feeding regime}
  \item{mean_length_mm}{Mean total length in mm}
  \item{survival_count}{Number of surviving larvae}
  \item{temp_c}{Average daily water temperature in Celsius}
}
}
\usage{
potteri_larvae
}
\description{
A dataset containing growth and survival metrics for Potter's Angelfish.
}
\keyword{datasets}
