% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hatchery_data.R
\name{read_hatchery_data}
\alias{read_hatchery_data}
\title{Read and Tidy Hatchery Datasheet}
\usage{
read_hatchery_data(file_path)
}
\arguments{
\item{file_path}{Path to the file}
}
\value{
A tidy dataframe
}
\description{
Read and Tidy Hatchery Datasheet
}
