% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spawn_data}
\alias{spawn_data}
\title{Potter's Angelfish Spawning Data}
\format{
A data frame with the following columns:
\describe{
  \item{Date}{Date of collection}
  \item{Viable}{Count of viable eggs}
  \item{Unviable}{Count of unviable eggs}
  \item{Total}{Total eggs collected}
  \item{Viability_Percent}{Percentage of eggs that were viable}
}
}
\usage{
spawn_data
}
\description{
Historical spawning data for Centropyge potteri, recording daily egg production and viability.
}
\keyword{datasets}
