% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_potter}
\alias{theme_potter}
\title{Custom Theme for Potterful Plots}
\usage{
theme_potter(base_size = 12)
}
\arguments{
\item{base_size}{Base font size.}
}
\description{
Custom Theme for Potterful Plots
}
